pyproject.toml
[tool.poetry]
name = "drf_github_stars"
version = "0.1.0"
description = "using Django rest framework to retrieve highly starred github projects"
authors = ["Lauren Biggers <grumpylauren@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
django-filter = "^21.1"
djangorestframework = "^3.12.4"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
